+++ .docs/gen-help.ps1
@@ -16,6 +16,14 @@ foreach ($file in $files) {
    # to avoid sending any unwanted data down the pipeline.
    $null = $sb.Append("### [$($file.BaseName)]($($file.Name))`r`n`r`n")
    $null = $sb.Append("<!-- #include ""./synopsis/$($file.Name)"" -->`r`n`r`n")
+
+   # I found files where the name of the file did not match the top most
+   # title in the file. This will cause issues trying to load help for that
+   # function. So test that you can find # {FileName} in the file.
+   $stringToFind = "# $($file.BaseName)"
+   if($null -eq $(Get-ChildItem $($file.Name) | Select-String $stringToFind)) {
+      Write-Error "Title cannot be found in $($file.Name). Make sure the first header is # $($file.BaseName)`n$($File.Directory)\$File" -ErrorAction Stop
+   }
 }
 
 Set-Content -Path files.md -Value $sb.ToString()
@@ -33,7 +41,7 @@ if(-not (Get-Module platyPS -ListAvailable)) {
    Install-Module platyPS -Scope CurrentUser -Force
 }
 
-New-ExternalHelp ..\docs -OutputPath ..\en-US -Force
+New-ExternalHelp ..\docs -OutputPath ..\Source\en-US -Force
 
 # Run again and strip header
 Write-Output 'Cleaning doc files for publishing'