+++ Source/Public/Set-VSTeamAPIVersion.ps1
@@ -6,7 +6,7 @@ function Set-VSTeamAPIVersion {
       [string] $Target = 'TFS2017',
 
       [parameter(ParameterSetName = 'Service', Mandatory = $true, Position = 0)]
-      [ValidateSet('Build', 'Release', 'Core', 'Git', 'DistributedTask', 'Tfvc', 'Packaging', 'MemberEntitlementManagement', 'ExtensionsManagement', 'ServiceFabricEndpoint')]
+      [ValidateSet('Build', 'Release', 'Core', 'Git', 'DistributedTask', 'Tfvc', 'Packaging', 'MemberEntitlementManagement', 'ExtensionsManagement', 'ServiceFabricEndpoint', 'Graph')]
       [string] $Service,
 
       [parameter(ParameterSetName = 'Service', Mandatory = $true, Position = 1)]
@@ -48,6 +48,9 @@ function Set-VSTeamAPIVersion {
             'ServiceFabricEndpoint' {
                [VSTeamVersions]::ServiceFabricEndpoint = $Version
             }
+            'Graph' {
+               [VSTeamVersions]::Graph = $Version
+            }
          }
       }
       else {
@@ -64,6 +67,7 @@ function Set-VSTeamAPIVersion {
                [VSTeamVersions]::MemberEntitlementManagement = ''
                [VSTeamVersions]::ServiceFabricEndpoint = '3.2'
                [VSTeamVersions]::ExtensionsManagement = '3.2-preview'
+               [VSTeamVersions]::Graph = ''
             }
             'TFS2017' {
                [VSTeamVersions]::Version = 'TFS2017'
@@ -77,19 +81,24 @@ function Set-VSTeamAPIVersion {
                [VSTeamVersions]::MemberEntitlementManagement = ''
                [VSTeamVersions]::ServiceFabricEndpoint = ''
                [VSTeamVersions]::ExtensionsManagement = '3.0-preview'
+               [VSTeamVersions]::Graph = ''
             }
             Default {
                [VSTeamVersions]::Version = $Target
-               [VSTeamVersions]::Git = '4.0'
-               [VSTeamVersions]::Core = '4.0'
-               [VSTeamVersions]::Build = '4.0'
-               [VSTeamVersions]::Release = '4.1-preview'
-               [VSTeamVersions]::DistributedTask = '4.1-preview'
-               [VSTeamVersions]::Tfvc = '4.0'
-               [VSTeamVersions]::Packaging = '4.0-preview'
-               [VSTeamVersions]::MemberEntitlementManagement = '4.1-preview'
-               [VSTeamVersions]::ServiceFabricEndpoint = '4.1-preview'
-               [VSTeamVersions]::ExtensionsManagement = '4.1-preview.1'
+               [VSTeamVersions]::Git = '5.1-preview.1'
+               [VSTeamVersions]::Core = '5.1-preview.1'
+               [VSTeamVersions]::Build = '5.0'
+               [VSTeamVersions]::Release = '5.1-preview.1'
+               [VSTeamVersions]::DistributedTask = '5.0-preview.2'
+               [VSTeamVersions]::Tfvc = '5.0'
+               [VSTeamVersions]::Packaging = '5.1-preview.1'
+               [VSTeamVersions]::MemberEntitlementManagement = '5.1-preview.1'
+               # This version is never passed to the API but is used to evaluate
+               # if Service Fabric is enabled for the account. Just set it to 
+               # match Distributed Task for AzD
+               [VSTeamVersions]::ServiceFabricEndpoint = '5.0-preview.2'
+               [VSTeamVersions]::ExtensionsManagement = '5.1-preview.1'
+               [VSTeamVersions]::Graph = '5.1-preview.1'
             }
          }
       }
@@ -106,4 +115,5 @@ function Set-VSTeamAPIVersion {
    Write-Verbose "MemberEntitlementManagement: $([VSTeamVersions]::MemberEntitlementManagement)"
    Write-Verbose "ServiceFabricEndpoint: $([VSTeamVersions]::ServiceFabricEndpoint)"
    Write-Verbose "ExtensionsManagement: $([VSTeamVersions]::ExtensionsManagement)"
+   Write-Verbose "Graph: $([VSTeamVersions]::Graph)"
 }